Added DTOs and functionality for the NFTs
