+++ src/infrastructure/Workers/CoinIndexer/CoinIndexerEvents.ts
@@ -1,5 +1,7 @@
 import { CoinRecord, CoinSpend } from '@dignetwork/chia-block-listener';
 import { Block } from '../../../application/entities/Block';
+import { Nft } from './Nft';
+import { Cat } from './AssetCats';
 
 
 export enum CoinIndexerEventNames {
@@ -24,11 +26,11 @@ export interface CoinIndexerEvents {
   on(event: CoinIndexerEventNames.NewBlockIngested, listener: (event: Block) => void): this;
   emit(event: CoinIndexerEventNames.NewBlockIngested, eventData: Block): boolean;
 
-  on(event: CoinIndexerEventNames.NftSpend, listener: (event: any) => void): this;
-  emit(event: CoinIndexerEventNames.NftSpend, eventData: any): boolean;
+  on(event: CoinIndexerEventNames.NftSpend, listener: (event: Nft) => void): this;
+  emit(event: CoinIndexerEventNames.NftSpend, eventData: Nft): boolean;
 
-  on(event: CoinIndexerEventNames.CatSpend, listener: (event: any) => void): this;
-  emit(event: CoinIndexerEventNames.CatSpend, eventData: any): boolean;
+  on(event: CoinIndexerEventNames.CatSpend, listener: (event: Cat) => void): this;
+  emit(event: CoinIndexerEventNames.CatSpend, eventData: Cat): boolean;
 
   on(event: CoinIndexerEventNames.DidSpend, listener: (event: any) => void): this;
   emit(event: CoinIndexerEventNames.DidSpend, eventData: any): boolean;
